// src/interfaces.ts
export interface IFeature {
	heading: string;
	iconName: string;
	imageUrl: string;
	link: string;
	content: string;
}

export interface IShowcase {
	heading: string;
	iconName: string;
	imageUrl: string;
	link: string;
	content: string;
}

export interface IExpertise {
	title: string;
	icon: string;
	body: string;
}

export interface ICustomer {
	comment: string;
	company: string;
	name: string;
	photo: string;
}